move default text form field
